tests: Moved un-atomic "unit" test to become a integration test
