+++ src/test/java/de/dhbw/karlsruhe/students/mailflow/unit/FileMailboxRepositoryTest.java
@@ -7,10 +7,7 @@ import de.dhbw.karlsruhe.students.mailflow.core.domain.email.exceptions.MailboxL
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.exceptions.MailboxSavingException;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.value_objects.Address;
 import de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email.parsing.FileMailboxRepository;
-import de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email.parsing.JSONMailboxConverter;
 import de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email.parsing.MailboxConverter;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -63,33 +60,6 @@ class FileMailboxRepositoryTest {
     Mailbox foundMailbox = fileMailboxRepository.findByAddressAndType(mailboxOwner, mailboxType);
     // Assert
     Assertions.assertEquals(searchingMailbox, foundMailbox);
-  }
-
-  @ParameterizedTest(name = "should create directories and files for mailboxType {0}")
-  @EnumSource(MailboxType.class)
-  /**
-   * #BugFix
-   */
-  void searchingNonExistentMailboxShouldCreateParsableAndEmptyMailbox(MailboxType mailboxType)
-      throws MailboxSavingException, MailboxLoadingException {
-    // Arrange
-    Address mailboxOwner = new Address("someOwner", "someDomain.de");
-    Mailbox searchingMailbox = Mailbox.create(mailboxOwner, Map.of(), mailboxType);
-    File directoryOfUser = new File(allMailboxesDirectory, mailboxOwner.toString());
-    File mailboxFile = new File(directoryOfUser, mailboxType.getStoringName() + ".json");
-
-    this.fileMailboxRepository =
-        new FileMailboxRepository(new JSONMailboxConverter(), allMailboxesDirectory);
-
-    // Act
-    Mailbox foundMailbox = fileMailboxRepository.findByAddressAndType(mailboxOwner, mailboxType);
-
-    // Assert
-    Assertions.assertTrue(mailboxFile.exists());
-
-    assertEquals(foundMailbox.getOwner(), searchingMailbox.getOwner());
-    assertEquals(foundMailbox.getEmailList(), searchingMailbox.getEmailList());
-    assertTrue(foundMailbox.getEmailList().isEmpty());
   }
 
   @ParameterizedTest(name = "should create file for mailboxType {0}")
